Don't recommend Terminus to run composer update (#120)
As most users tend to run out of memory when running `terminus composer site.env -- update`, we should just stop telling them to do it. This is almost guaranteed to run out of memory. We should choose a single supported way to do this as adoption is growing and it is creating confusion.